Move slack commands in run
