+++ test/unit/run/services/slack/commands_test.js
@@ -5,6 +5,7 @@ const {
   createAndDeployPixLCMS,
   createAndDeployPixUI,
   createAndDeployPixSiteRelease,
+  createAndDeployPixDatawarehouse,
 } = require('../../../../../run/services/slack/commands');
 const releasesServices = require('../../../../../common/services/releases');
 const githubServices = require('../../../../../common/services/github');
@@ -121,4 +122,27 @@ describe('Services | Slack | Commands', () => {
     });
   });
 
+  describe('#createAndDeployPixDatawarehouse', () => {
+    beforeEach(async () => {
+      // given
+      const payload = { text: 'minor' };
+      // when
+      await createAndDeployPixDatawarehouse(payload);
+    });
+
+    it('should publish a new release', () => {
+      // then
+      sinon.assert.calledWith(releasesServices.publishPixRepo, 'pix-db-replication', 'minor');
+    });
+
+    it('should retrieve the last release tag from GitHub', () => {
+      // then
+      sinon.assert.calledWith(githubServices.getLatestReleaseTag, 'pix-db-replication');
+    });
+
+    it('should deploy the release', () => {
+      // then
+      sinon.assert.calledWith(releasesServices.deployPixRepo);
+    });
+  });
 });
